+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0-pre-8
+* Fix Events Global Error `onError`
+* Modify `listen` to action events
+
 ## 1.0.0-pre-7
 * modify `ActionRequestError` catch and handle errors
 * modify `execute` to handle response with errors